+++ tests/meltano/core/test_project_settings_service.py
@@ -72,3 +72,171 @@ class TestProjectSettingsService:
             )
 
             assert_value_source("from_env", SettingValueSource.ENV)
+
+    def test_path_prefix(self, session, subject):
+        assert not subject.path_prefix
+        assert not subject.name_prefix
+
+        subject.set("project_id", "from_dotenv", store=SettingValueStore.DOTENV)
+        subject.set("project_id", "from_yml", store=SettingValueStore.MELTANO_YML)
+        subject.set(
+            "project_id", "from_db", store=SettingValueStore.DB, session=session
+        )
+
+        subject.set("ui.server_name", "from_yml", store=SettingValueStore.MELTANO_YML)
+        subject.set("ui.custom", "from_yml", store=SettingValueStore.MELTANO_YML)
+        subject.set("ui.custom", "from_dotenv", store=SettingValueStore.DOTENV)
+
+        assert subject.get_with_source("project_id") == (
+            "from_dotenv",
+            SettingValueSource.DOTENV,
+        )
+
+        subject.path_prefix = ["ui"]
+        assert subject.name_prefix == "ui."
+
+        # definitions
+        definition_names = [setting_def.name for setting_def in subject.definitions()]
+        assert all(setting_name.startswith("ui.") for setting_name in definition_names)
+        assert "ui.server_name" in definition_names
+        assert "ui.custom" in definition_names
+
+        # config_with_metadata
+        config_keys = subject.config_with_metadata().keys()
+        assert all(setting_name.startswith("ui.") for setting_name in config_keys)
+        assert "ui.server_name" in config_keys
+        assert "ui.custom" in config_keys
+
+        # as_dict
+        dict_keys = subject.as_dict().keys()
+        assert all(setting_name.startswith("ui.") for setting_name in dict_keys)
+        assert "ui.server_name" in dict_keys
+        assert "ui.custom" in dict_keys
+
+        # as_env
+        env_keys = subject.as_env().keys()
+        assert all(setting_name.startswith("MELTANO_UI_") for setting_name in env_keys)
+        assert "MELTANO_UI_SERVER_NAME" in env_keys
+        assert "MELTANO_UI_CUSTOM" in env_keys
+
+        # get_with_source
+        assert subject.get_with_source("project_id") == (
+            None,
+            SettingValueSource.DEFAULT,
+        )
+        assert subject.get_with_source("server_name") == (
+            "from_yml",
+            SettingValueSource.MELTANO_YML,
+        )
+        assert subject.get_with_source("custom") == (
+            "from_dotenv",
+            SettingValueSource.DOTENV,
+        )
+
+        # get
+        assert subject.get("server_name") == "from_yml"
+        assert subject.get("custom") == "from_dotenv"
+
+        # set
+        subject.set("server_name", "from_dotenv", store=SettingValueStore.DOTENV)
+
+        assert subject.get_with_source("server_name") == (
+            "from_dotenv",
+            SettingValueSource.DOTENV,
+        )
+
+        subject.set(
+            "server_name", "from_db", store=SettingValueStore.DB, session=session
+        )
+
+        # unset
+        subject.unset("server_name", store=SettingValueStore.DOTENV)
+
+        assert subject.get_with_source("server_name") == (
+            "from_yml",
+            SettingValueSource.MELTANO_YML,
+        )
+
+        subject.unset("server_name", store=SettingValueStore.MELTANO_YML)
+
+        assert subject.get_with_source("server_name", session=session) == (
+            "from_db",
+            SettingValueSource.DB,
+        )
+
+        subject.unset("server_name", store=SettingValueStore.DB, session=session)
+
+        assert subject.get_with_source("server_name", session=session) == (
+            None,
+            SettingValueSource.DEFAULT,
+        )
+
+        # reset
+        subject.set("server_name", "from_dotenv", store=SettingValueStore.DOTENV)
+        subject.set("server_name", "from_yml", store=SettingValueStore.MELTANO_YML)
+        subject.set(
+            "server_name", "from_db", store=SettingValueStore.DB, session=session
+        )
+
+        assert subject.get_with_source("server_name") == (
+            "from_dotenv",
+            SettingValueSource.DOTENV,
+        )
+        assert subject.get_with_source("custom") == (
+            "from_dotenv",
+            SettingValueSource.DOTENV,
+        )
+
+        subject.reset(store=SettingValueStore.DOTENV)
+
+        assert subject.get_with_source("server_name") == (
+            "from_yml",
+            SettingValueSource.MELTANO_YML,
+        )
+        assert subject.get_with_source("custom") == (
+            "from_yml",
+            SettingValueSource.MELTANO_YML,
+        )
+
+        subject.reset(store=SettingValueStore.MELTANO_YML)
+
+        assert subject.get_with_source("server_name", session=session) == (
+            "from_db",
+            SettingValueSource.DB,
+        )
+        assert subject.get_with_source("custom") == (None, SettingValueSource.DEFAULT)
+
+        subject.reset(store=SettingValueStore.DB, session=session)
+
+        assert subject.get_with_source("server_name", session=session) == (
+            None,
+            SettingValueSource.DEFAULT,
+        )
+
+        subject.path_prefix = []
+
+        assert subject.get_with_source("project_id") == (
+            "from_dotenv",
+            SettingValueSource.DOTENV,
+        )
+
+        subject.reset(store=SettingValueStore.DOTENV)
+
+        assert subject.get_with_source("project_id") == (
+            "from_yml",
+            SettingValueSource.MELTANO_YML,
+        )
+
+        subject.reset(store=SettingValueStore.MELTANO_YML)
+
+        assert subject.get_with_source("project_id", session=session) == (
+            "from_db",
+            SettingValueSource.DB,
+        )
+
+        subject.reset(store=SettingValueStore.DB, session=session)
+
+        assert subject.get_with_source("project_id", session=session) == (
+            None,
+            SettingValueSource.DEFAULT,
+        )