+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 - [#2100](https://gitlab.com/meltano/meltano/-/issues/2100) Let settings be stored directly in `.env` using `meltano config set --store=dotenv`
 - [#2099](https://gitlab.com/meltano/meltano/-/issues/2099) Let Meltano settings be configured using `meltano config meltano {list,set,unset,reset}`
+- [#2099](https://gitlab.com/meltano/meltano/-/issues/2099) Let namespaced (nested) Meltano settings be configured using `meltano config <namespace>`, e.g. `meltano config ui set server_name meltano.example.com` as opposed to `meltano config meltano set ui server_name meltano.example.com`
 
 ### Changes